# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CyberLearn LMS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /certificates/generate:
    post:
      operationId: generateCertificate
      tags: [certificates]
      summary: Generate certificate (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CertificateInput"
      responses:
        "201":
          description: Certificate generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Certificate"
  /certificates/{hash}/verify:
    get:
      operationId: verifyCertificate
      tags: [certificates]
      summary: Verify certificate by hash (public)
      parameters:
        - name: hash
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Certificate verification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CertificateVerification"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── NOTIFICATIONS ─────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ─── AUTH SCHEMAS ─────────────────────────────────────────────────────────
...
    Certificate:
      type: object
      required: [id, userId, courseId, issuedAt, certHash]
      properties:
        id:
          type: integer
        userId:
          type: integer
        courseId:
          type: integer
        issuedAt:
          type: string
        certHash:
          type: string
        pdfUrl:
          type: ["string", "null"]
        courseName:
          type: string
        studentName:
          type: string
    CertificateInput:
      type: object
      required: [userId, courseId]
      properties:
        userId:
          type: integer
        courseId:
          type: integer
    CertificateVerification:
      type: object
      required: [valid, certificate]
      properties:
        valid:
          type: boolean
        certificate:
          $ref: "#/components/schemas/Certificate"

    # ─── NOTIFICATION SCHEMAS ─────────────────────────────────────────────────
...